moved team specific components to their own directory
