+++ frontend/src/components/team/member_card.vue
