+++ frontend/src/views/team/settings.vue
@@ -89,7 +89,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
-import TeamNav from '@/components/team_nav.vue'
+import TeamNav from '@/components/team/nav.vue'
 import DeleteTeam from '@/components/modals/confirmations/delete_team.vue'
 import { TeamUpdateErrors } from '@/interfaces/responses'
 import Team from '@/interfaces/team'
